pyproject.toml
[tool.poetry]
name = "classic-rock-hits-cli"
version = "0.1.0"
description = "A CLI application that provides information about classic rock hits for a given year."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "classic_rock_hits_cli"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
classic-rock-hits = "classic_rock_hits_cli.main:main"